requirements.txt
amqp==5.2.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
black==24.4.2
blessed==1.20.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-debug-toolbar==4.4.2
django-filter==24.2
django-picklefield==3.2
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gevent==24.2.1
greenlet==3.0.3
idna==3.7
jinxed==1.2.1
kombu==5.3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pycparser==2.22
PyJWT==2.8.0
pyTelegramBotAPI==4.20.0
python-baseconv==1.2.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==4.5.4
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stripe==10.2.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.4.post2
